提交

+++ app/Services/Impl/SpecsServiceImpl.php
@@ -72,6 +72,9 @@ class SpecsServiceImpl implements SpecsService
         if (!$specs) {
             throw new \Exception('规格不存在');
         }
+        if (!$specs->delete_time != 0) {
+            throw new \Exception('规格已删除');
+        }
         unset($fields['id']);
         $this->model->updateById($id, $fields);
     }
@@ -81,6 +84,9 @@ class SpecsServiceImpl implements SpecsService
         $specs = $this->model->find($id);
         if (!$specs) {
             throw new \Exception('规格不存在');
+        }
+        if (!$specs->delete_time != 0) {
+            throw new \Exception('规格已删除');
         }
         $this->model->deleteById($id);
     }
